requirements.txt
# core dashboard + viz
streamlit>=1.36.0,<2.0.0
plotly>=5.20.0,<6.0.0
matplotlib>=3.7.0,<4.0.0

# numeric / data stack
numpy>=1.26.0,<2.0.0
pandas>=2.0.0,<2.2.0
pyarrow>=14.0.2,<18.0.0
scikit-learn>=1.3.0,<1.5.0


# Nixtla ecosystem (stat + ML + neural)
statsforecast>=1.7.3,<3.0.0
mlforecast>=0.10.0,<2.0.0
neuralforecast>=1.7.2,<3.0.0

# trading SDK
alpaca-py>=0.20.0,<1.0.0

s5-pytorch>=0.2.0
# Heavy optional RL deps moved to rl_requirements.txt to keep resolver light
# Install your CUDA-enabled torch build separately, e.g. (Windows CUDA 12.1):
# pip install --index-url https://download.pytorch.org/whl/cu121 torch==2.1.2+cu121 torchvision==0.16.2+cu121 torchaudio==2.1.2+cu121
